Export receivers in GsmCellBroadcastHandler
The runtime receivers in GsmCellBroadcastHandler are currently
registered with the RECEIVER_NOT_EXPORTED flag; this prevents
other apps from being able to send broadcasts to these receivers.
However, these receivers are guarded by a signature permission,
and broadcasts are expected from external apps that are signed
with the same signing key. This commit registers these receivers
with the RECEIVER_EXPORTED flag to allow the expected broadcasts
from external apps that are signed with the same signing key and
granted the signature permission that is guarding these receivers.

Fixes: 245018067
Test: atest GsmCellBroadcastHandlerTest
Change-Id: I1ff3dd9deac6faa017b5f856311b1aaf686ba875