Suppress lint for methods using APIs promoted from system to public
Android 14 added the ability for a broadcasting app to share its
identity with a receiver. System APIs already existed to send a
broadcast with a BroadcastOptions Bundle, so these APIs were
promoted from system to public; however our tooling does not
currently recognize that these APIs were previously available to
modules that can invoke system APIs resulting in NewApi lint
errors. This commit annotates all of the methods that invoke one
of these newly promoted APIs to suppress the NewApi lint error;
this will have no affect on functionality since the APIs are still
available on previous platform releases as system APIs, and these
annotations will be removed when the tooling supports system to
public APIs.

Bug: 259743961
Test: errorprone build
Change-Id: I40e72bfa09da01188976b08ee909f5785449577a